pdf_to_image/requirements.txt
opencv-python==4.5.5.62
matplotlib==3.6.3
numpy==1.24.0
Pillow==9.2.0
pdf2image==1.16.3
